Day one complete!
This was pretty simple, and itertools did all the heavy lifting for me
with its wonderful "tuple_windows" method. I also set up what I hope
is a sufficiently generic set of glue code, that I'll be able to use
to solve the next 24 days (or however far I get).