Day three complete!
There was a bit of code duplication in part 2, that I'm not a fan of.
However, I'm trying to move forward instead of spending ages
perfecting it, no matter how tempting that is. There probably would
have been similar duplication for part 1, had I not found that neat
XOR trick.