Day four complete!
This one really stretched my rust skills, and led to some googling.
In order to call ".collect" and get a fixed size array out of it,
with dropping and error-handling in case too many or too few elements
were supplied, I needed to get into:

* Unsafe MaybeUninit operations
* Unsafe pointer operations to replicate currently unstable
  MaybeUninit methods
* Const generics (first time using these for me)

However, it did make my "parse a bingo board" function incredibly
slick. And it's saved as a utility that should prove more useful than
the parser that I wrote for day 2's input format.