Day 6 complete!
This time the input was comma separated, instead of line separated,
so I copied and pasted some utility code. Maybe should find a way to
share code here, but for now it's working, so I'll leave it.

Also, fun algorithm problem! Super easy to write out the linear
solution I used, but only because I'd dealt with something like this
in a previous year. Also, the way the example is laid out kinda pushes
you toward thinking about it on a fish-by-fish basis, which could
have led to really slow code if I wasn't already planning something
like this.