Day 9 done, but at what cost?
This one was brutal, and mostly because I took the complete wrong
approach implementing it for the first time. Anyway, some refactoring
later and it's looking okay. So what went wrong?

Firstly, I didn't build a nice abstraction around finding neighbours.
Finding neighbours, and accounting for out-of-bounds cases, is
something that got messy pretty quickly in rust, due to the fact that
you're forced to handle it at the type level, rather than just with
normal bounds checks. ".checked_sub" anyone? But that's not the end
of the world - messy code still gets points in challenges like this,
and I make a point to refactor before I upload the code anyway. So
what really went wrong?

I missed an algorithm thing, I guess. The core of part two (the way I
implemented it, at least) is finding a "canonical" representation of
a basin that can be easily hashed, and finding a method to turn a
point in the basin into its canonical representation. Once you've got
that, it's a simple matter of looping through the points, filling up
a "which basin is biggest" HashMap, and then looping through that to
find the largest basins. The tricky bit is that canonical
representation...

In the solution I've committed, the canonical representation is just
the coordinates of the low point of the basin. I realised overnight
that you can just walk down a lava flow to reach the lowest point
(and that you wouldn't even have to do the colouring or branching
that most graph algorithms involve). However, the version I committed
was a lot worse.

https://en.wikipedia.org/wiki/Disjoint-set_data_structure

That's what I did. I remember running into these in my second year of
uni, and I basically never forgot about them. My "canonical
representation of a basin" became the set representative, and I just
walked through the map joining sets together left right and centre.
In theory, it wasn't a bad solution. If it wasn't for the fact that
every basin is guaranteed to have a single low point that's easy to
reach, it would probably be the optimal way to solve this (with some
tidying up). But it turns out remembering how these data structures
work and implementing an algorithm with one in practice are two very
different things. Who would have guessed? Anyway, here are some of
the things that went wrong.

1. Accidentally made two elements each-others' representatives. Boom!
Stack Overflow! And it turns out, stack overflows kill whatever
mechanism flushes stdout on exit, so I couldn't see my debug print
statements either.

2. Forgot to re-flatten paths when I was summing up how big each
basin was at the end. Turns out that set representatives are only a
canonical representation if you flatten them out - otherwise they're
just nodes halfway down a tree. And so my basins didn't join up well
enough, meaning I had too many basins that were too small.

3. Just as a debugging thing, set representatives aren't great to look
at when figuring out where your bugs are. Unlike the low point of a
basin, there is no real way for me to look at a map of the inputs and
figure out where the set representative is meant to be. Which means
there's no real way for me to look at a printout of where the set
representative is, and tell whether it's correct or not.