Day 13 complete!
I've noticed that the parsing code is getting more and more verbose as
we go through the month. I think I'm going to introduce some extra
utilities, just to make things cleaner, and do a bit of a refactor
before the next day.

Also, I think the hardest (or at least most doubt-inducing) part of
today's challenge was reading the output for part 2. I'm not good at
Captchas on the best of days.