Day 14 done!
This one was tricky. I mean, the final solution was easy enough to
write once I started thinking in digrams. And the original solution
was a little messy at first, and took some refactoring, but all in
all wasn't too bad.

...but it was tricky.

My first solution just looped 40 times, re-calculating the string. By
the time I shut it down it was using all the ram on my computer, and
all the free space on my disk as swap. And it was around 30-31 loops
in.

My second solution is included in the actual code I've committed, as
the solution for part 1, because I'm really proud of it. It uses lazy
iterators to build an expression for calculating the final string
all-at-once, without any intermediate steps - and **then** it walks
over that iterator. This gives us O(1) memory usage, and it gives
LLVM the opportunity to work its magic in really freaky ways. As I
commit this, that solution has churned through 274 gigabytes of data
or more, and is still running on a cool 628Kb of memory usage.

The final brute force solution just obfuscated a bunch of things
pointlessly, to try to gain some split-second performance
improvements. It was a little faster, but not more than 1% or so, and
the code was more than a little sketchy-looking.

...and then I realised I could work this with digrams. I tried for a
while staring at a spreadsheet, seeing if I could find (or make) a
pattern in how the number of "B"s increased over time, before
realising that the real trick is counting how many "AB"s there are
(next to each-other) in the string. From there it was honestly pretty
clear sailing.

Damn that took some thinking of though